reopen issues possible

+++ src/modules/MUBoard/lib/MUBoard/Util/Model.php
@@ -18,26 +18,54 @@ class MUBoard_Util_Model extends MUBoard_Util_Base_Model
 {
 
 	/**
-	 * This function counts the call of a posting
-	 * @param id   id of the ticket to close
+	 * This function closes a posting
+	 * @param id   id of the posting to close
 	 */
 
 	public static function closePosting($id) {
 
-		// build posting repository
-		$repository = MUBoard_Util_Model::getPostingRepository();
+		// Security ckeck
+		if (SecurityUtil::checkPermission('MUBoard::', '::', ACCESS_ADMIN)) {
 
-		$entity = $repository->selectById($id);
+			// build posting repository
+			$repository = MUBoard_Util_Model::getPostingRepository();
 
-		$serviceManager = ServiceUtil::getManager();
-		$entityManager = $serviceManager->getService('doctrine.entitymanager');
+			$entity = $repository->selectById($id);
 
-		$entity->setState(0);
+			$serviceManager = ServiceUtil::getManager();
+			$entityManager = $serviceManager->getService('doctrine.entitymanager');
 
-		$entityManager->flush();
+			$entity->setState(0);
 
-		return true;
+			$entityManager->flush();
 
+			return true;
+		}
+	}
+
+	/**
+	 * This function reopens a posting
+	 * @param id   id of the posting to open
+	 */
+
+	public static function openPosting($id) {
+
+		if (SecurityUtil::checkPermission('MUBoard::', '::', ACCESS_ADMIN)) {
+
+			// build posting repository
+			$repository = MUBoard_Util_Model::getPostingRepository();
+
+			$entity = $repository->selectById($id);
+
+			$serviceManager = ServiceUtil::getManager();
+			$entityManager = $serviceManager->getService('doctrine.entitymanager');
+
+			$entity->setState(1);
+
+			$entityManager->flush();
+
+			return true;
+		}
 	}
 
 	/**
@@ -119,40 +147,44 @@ class MUBoard_Util_Model extends MUBoard_Util_Base_Model
 
 	public static function quitAbo($posting = 0, $forum = 0, $category = 0) {
 
-		// build posting repository
-		$repository = MUBoard_Util_Model::getAboRepository();
-		// get actual user id
-		$userid = UserUtil::getVar('uid');
-		// quit abo of posting if posting is taken
-		if ($posting > 0) {
-			// look for abo
-			$where = 'tbl.postingid = \'' . DataUtil::formatForStore($posting) . '\'';
-			$where .= ' AND ';
-			$where .= 'tbl.userid = \'' . DataUtil::formatForStore($userid) . '\'';
-			$abos = $repository->selectWhere($where);
-			$serviceManager = ServiceUtil::getManager();
-			$entityManager = $serviceManager->getService('doctrine.entitymanager');
-			foreach ($abos as $abo) {
-				$entityManager->remove($abo);
-				$entityManager->flush();
-			}
-		}
+		// Security ckeck
+		if (SecurityUtil::checkPermission('MUBoard::', '::', ACCESS_COMMENT)) {
 
-		if ($forum > 0) {
-			// look for abo
-			$where = 'tbl.forumid = \'' . DataUtil::formatForStore($forum) . '\'';
-			$where .= ' AND ';
-			$where .= 'tbl.userid = \'' . DataUtil::formatForStore($userid) . '\'';
-			$abos = $repository->selectWhere($where);
-			$serviceManager = ServiceUtil::getManager();
-			$entityManager = $serviceManager->getService('doctrine.entitymanager');
-			foreach ($abos as $abo) {
-				$entityManager->remove($abo);
-				$entityManager->flush();
+			// build posting repository
+			$repository = MUBoard_Util_Model::getAboRepository();
+			// get actual user id
+			$userid = UserUtil::getVar('uid');
+			// quit abo of posting if posting is taken
+			if ($posting > 0) {
+				// look for abo
+				$where = 'tbl.postingid = \'' . DataUtil::formatForStore($posting) . '\'';
+				$where .= ' AND ';
+				$where .= 'tbl.userid = \'' . DataUtil::formatForStore($userid) . '\'';
+				$abos = $repository->selectWhere($where);
+				$serviceManager = ServiceUtil::getManager();
+				$entityManager = $serviceManager->getService('doctrine.entitymanager');
+				foreach ($abos as $abo) {
+					$entityManager->remove($abo);
+					$entityManager->flush();
+				}
 			}
-		}
 
-		return true;
+			if ($forum > 0) {
+				// look for abo
+				$where = 'tbl.forumid = \'' . DataUtil::formatForStore($forum) . '\'';
+				$where .= ' AND ';
+				$where .= 'tbl.userid = \'' . DataUtil::formatForStore($userid) . '\'';
+				$abos = $repository->selectWhere($where);
+				$serviceManager = ServiceUtil::getManager();
+				$entityManager = $serviceManager->getService('doctrine.entitymanager');
+				foreach ($abos as $abo) {
+					$entityManager->remove($abo);
+					$entityManager->flush();
+				}
+			}
+
+			return true;
+		}
 	}
 
 	/**
